Refactor architecture to break circular dependencies and split God Object
